let's give it a go

+++ zappa/cli.py
@@ -1696,7 +1696,7 @@ class ZappaCLI(object):
             try:
                 if os.path.isfile(self.zip_path):
                     os.remove(self.zip_path)
-                if os.path.isfile(self.handler_path):
+                if self.handler_path and os.path.isfile(self.handler_path):
                     os.remove(self.handler_path)
             except Exception as e: # pragma: no cover
                 sys.exit(-1)
